requirements.txt
discord-py-interactions==5.9.2
python-dotenv==1.0.0
pytz==2023.3.post1